move tests to test_vm
